move scrape away from api route
